+++ tikv-client-common/Cargo.toml
@@ -8,7 +8,6 @@ edition = "2018"
 failure = "0.1"
 grpcio = { version = "0.5.0", features = [ "secure", "prost-codec" ], default-features = false }
 futures = { version = "0.3.1", features = ["compat", "async-await", "thread-pool"] }
-kvproto = { git = "https://github.com/pingcap/kvproto.git", features = [ "prost-codec" ], default-features = false }
 lazy_static = "1"
 regex = "1"
 serde = "1.0"
@@ -16,6 +15,8 @@ serde_derive = "1.0"
 log = "0.4"
 proptest = "0.9"
 proptest-derive = "0.1.0"
+derive-new = "0.5"
+kvproto = { git = "https://github.com/pingcap/kvproto.git", features = [ "prost-codec" ], default-features = false }
 
 [dependencies.prometheus]
 version = "0.8"
